requirements.txt
openai>=1.0.0
reportlab>=3.6.12
python-dotenv>=1.0.0
requests>=2.28.0
Pillow>=9.0.0
pyinstaller